+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Minimal Archaeology
 ## Latest Version
 
+### 9.0.1
+
+- Added the option to lock the Companion in place, disabling dragging
+- Added the option to persist Companion position on all characters using the same settings profile
+
 ### 9.0.0
 
 - New feataure: **Companion** (**distance tracker** and more, details down below)